util: split out logging component into overlay and expandable flavors.
