--- lib/elixir_rpg_web/components/logger_component.ex
@@ -1,35 +1,0 @@
-defmodule ElixirRpgWeb.LogOverlayComponent do
-  use ElixirRpgWeb, :live_component
-
-  #  <div class="fixed bottom-4 right-4">
-  #  <details class="bg-white shadow-lg rounded-lg overflow-hidden">
-  #    <summary class="bg-gray-800 text-white p-2 cursor-pointer">
-  #      <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6 inline-block mr-2" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-  #        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M13 10V3L4 14h7v7l9-11h-7z" />
-  #      </svg>
-  #      Logs
-  #    </summary>
-  #    <div class="p-4 max-h-96 overflow-y-auto">
-  #      <ul id="log-messages" phx-update="stream" class="text-sm">
-  #        <li :for={{dom_id, message} <- @streams.log_messages} id={dom_id} class="mb-1 pb-1 border-b border-gray-200">
-  #          <%= message.message %>
-  #        </li>
-  #      </ul>
-  #    </div>
-  #  </details>
-  #</div>
-
-  def render(assigns) do
-    ~H"""
-    <div class="fixed inset-0 flex items-center justify-center pointer-events-none">
-      <div class="bg-black bg-opacity-50 p-4 rounded-lg w-2/3 h-1/2 overflow-y-auto">
-        <ul id="log-messages" phx-update="stream" class="text-white">
-          <li :for={{dom_id, message} <- @streams.log_messages} id={dom_id}>
-            <%= message.message %>
-          </li>
-        </ul>
-      </div>
-    </div>
-  """
-  end
-end
